Metronomes Data

Measurements of five models of metronome: Neewer, Neewer (photo), Patent, TB 06 and TB 07. The first two correspond to a contemporary metronome, a Neewer NW-707. The difference is that Neewer corresponds to precise direct measurements in a dismantled metronome, while Neewer (photo) measurements were taken from a photograph. The rest of the models follow this methodology. The third one, Patent, corresponds to the original diagram from Maelzel's 1815 English patent. Finally, TB 06 and TB 07 are metronomes number 6 and 7, respectively, from Tony Bingham's collection (see the references).

metr.neewer

metr.marks

metr.params

Format

metr.neewer provides oscillation frequency measurements for the Neewer metronome.

mark	metronome mark, in pulses per minute.

bpm	measured frequency, in pulses per minute.

bpm.se	standard error of bpm, in the same units.

metr.marks provides position measurements of each metronome mark for all the metronomes.

model	metronome model.

mark	metronome mark, in pulses per minute.

r	distance from the shaft to the mark, in mm.

r.se	standard error of r, in the same units.

metr.params provides measurements of fixed dimensions for all the metronomes. All the additional columns suffixed with .se are the standard errors of the corresponding variables in the same units.

model	metronome model.

rcm	position of the center of mass of the upper mass with respect to the metronome mark, in mm.

l	length of the rod over the shaft, in mm.

R	distance from the shaft to the center of mass of the lower mass, in mm.

L	length of the rod below the shaft, in mm.

A	oscillation amplitude, in degrees.

References

Bingham, T. and Turner, A. (2017) Metronomes and Musical Time: Catalogue of the Tony Bingham Collection at the exhibition AUF TAKT! Held in the Museum für Musik, Basel. London. ISBN: 978-0-946113-11-8.