Beethoven's Symphonies Data

Six data sets with tempo measurements plus additional information about Beethoven's nine symphonies.

Data sets sym.marks, sym.recordings and sym.duration provide detailed information about Beethoven's annotations as well as the recordings selected for this study (the complete symphonies from 36 different conductors).

Data sets sym.window, and sym.sample provide tempo measurements extracted using the Marsyas framework (and the tempo estimation algorithm by Percival and Tzanetakis; see the references) and distinct methodologies.

sym.marks

sym.recordings

sym.duration

sym.window

sym.sample

Format

sym.marks provides complete timing information extracted from Beethoven's scores: character annotations, tempo markings, beats and bars for the nine symphonies. Note that some movements have several character annotations (more than one section).

symphony	symphony number.

movement	movement number.

character	character annotation in this section.

mark	metronome mark.

beat	beats per bar.

bar.tot	total number of bars in this section.

bar.rep	total number of repeated bars in this section.

bar.1	total number of bars in first repetition boxes.

tsig	time signature, as expected by tmp_rectify_tsig

section	convenience section identifier (one per character).

Therefore, the theoretical duration, in minutes, without repetitions for each section can be computed as follows: (bar.tot - bar.1) * beat / mark. And the theoretical duration, in minutes, with repetitions is just the previous value + bar.rep * beat / mark.

sym.recordings provides complete information about the albums analyzed in this work (the nine symphonies for 36 different conductors).

conductor	conductor's name.

orchestra	orchestra(s)'s name.

title	album title.

label	album label.

upc	Universal Product Code.

date	recording dates.

year	release date.

ptype	performance type: romantic, Historically Informed (HI) or under HI influence.

sym.duration provides performed duration (track length) for the nine symphonies and 36 different conductors.

symphony	symphony number.

movement	movement number.

conductor	conductor's name.

duration	track length, in seconds.

sym.window provides continuous tempo measurements per symphony, movement and conductor by means of a sliding window.

symphony	symphony number.

movement	movement number.

conductor	conductor's name.

n	sample index. Samples with n=0 correspond to an estimation for the entire track.

start	start time for the sliding window, in seconds.

duration	window length, in seconds.

tempo	estimated tempo.

sym.sample, in contrast to sym.window, provides a single tempo estimation for each symphony, movement and conductor. The sample was collected at the end of each track (the "coda" of the movement), where the tempo is arguably more stable. It has the same variables as sym.window, except for n. Instead, sym.sample contains a section column to identify the section from which the tempo was sampled. This section identifier corresponds to the homonymous one present in sym.marks.

References

Tzanetakis, G. and Cook, P. (2000) "MARSYAS: a framework for audio analysis." Organised Sound, 4(3):169-175

Percival, G. and Tzanetakis, G. (2013) "An effective, simple tempo estimation method based on self-similarity and regularity." IEEE International Conference on Acoustics, Speech and Signal Processing, 241-245.

Percival, G. and Tzanetakis, G. (2014) "Streamlined Tempo Estimation Based on Autocorrelation and Cross-correlation with Pulses." IEEE/ACM Trans. Audio, Speech and Lang. Proc., 22(12):1765-1776.