Metronome Model Fit

Fit metronome's nondimensionalized masses, M. and m. (see metr_model), from a set of measurements and parameters.

metr_fit(.data, params, by = "model")

metr_predict(.fit, by = "model")

Arguments

.data
measurements of distance to the shaft for each metronome mark. It must contain the variables model, mark and r, as, e.g., the metr.marks data set.

params
the rest of the dimensions: correction of the center of mass of the upper mass rcm, length of the rod above the shaft l, length of the rod below the shaft L, distance of the lower mass to the shaft R and oscillation amplitude A, as, e.g., the metr.params data set.

by
variable to group by; e.g., the metronome model, to fit several metronomes.

.fit
output from metr_fit.

Value

A grouped data frame with model, M. (and error), mu. (and error) and the fit object.